Move theme provider to root
